+++ src/handlers/item-operations-handlers.ts
@@ -16,8 +16,7 @@ import {
   TodoistAPIError,
 } from "../errors.js";
 import { extractArrayFromResponse } from "../utils/api-helpers.js";
-
-const SYNC_API_URL = "https://api.todoist.com/sync/v9";
+import { SYNC_API_URL } from "../utils/api-constants.js";
 
 function getApiToken(): string {
   const token = process.env.TODOIST_API_TOKEN;